Fix occasional CRC error
Debugged and fixed with Aron.  There was an occasionally occuring error
where if an object being written had just the right (i.e. wrong) number
of bytes remaining when it wrapped over a sector, the sector number
would be incremented twice (should just be once).  A sector would be
left blank and the tail of the object would be written to the following
one.  When the logger finally attempted to read the object, its tail
would be read as blank and deserialize would throw an exception.